Fix bug removing multiple bookcars and delete console.logs
